Removing integer assertions

+++ src/helpers.jl
@@ -49,6 +49,6 @@ function nu(L::FiniteCoxeterGroup,iso_plevis::Vector,all_plevis::Vector)
 	return nu_value
 end
 
-function dimension_XY(G::FiniteCoxeterGroup,genus::BigInt,puncture::BigInt)
+function dimension_XY(G::FiniteCoxeterGroup,genus,puncture)
 	return (2*genus-2+puncture)*dimension(G) + 2*(rank(G)-semisimplerank(G)) - puncture*rank(G)
 end
